Trasformazione Id da int a long, Eliminazione degli id dai costruttori in vista di passaggio a Db per persistenza dei dati

+++ src/main/java/it/unicam/cs/ids/piattaforma_agricola_locale/model/repository/ICertificazioneRepository.java
@@ -8,11 +8,11 @@ import java.util.List;
 
 public interface ICertificazioneRepository {
         void save(Certificazione certificazione);
-        Certificazione findById(int id);
+        Certificazione findById(Long id);
         List<Certificazione> findAll();
-        List<Certificazione> findByProdottoId(int idProdotto);
-        List<Certificazione> findByAziendaId(int idAzienda); // Assumendo che DatiAzienda abbia un ID
-        void deleteById(int id);
-        void deleteByProdottoId(int idProdotto); // Utile se un prodotto viene rimosso
-        void deleteByAziendaId(int idAzienda);   // Utile se un'azienda viene rimossa
+        List<Certificazione> findByProdottoId(Long idProdotto);
+        List<Certificazione> findByAziendaId(Long idAzienda); // Assumendo che DatiAzienda abbia un ID
+        void deleteById(Long id);
+        void deleteByProdottoId(Long idProdotto); // Utile se un prodotto viene rimosso
+        void deleteByAziendaId(Long idAzienda);   // Utile se un'azienda viene rimossa
 }
